Move interface commands to individual files
