+++ server/app.js
@@ -2,6 +2,7 @@ const express = require('express');
 const app = express();
 const cors = require('cors');
 const dbinit = require('./src/models/database.js');
+app.use(cors());
 require('./src/utils/routes.js')(app);
 
 // const testFuncs = require('./src/utils/studentInterface.js');
